Refactor stack navigators into dedicated modules
